+++ products/error_tracking/frontend/scenes/ErrorTrackingScene/tabs/insights/queries.ts
@@ -1,0 +1,96 @@
+import { urls } from 'scenes/urls'
+
+import { InsightVizNode, NodeKind, TrendsQuery } from '~/queries/schema/schema-general'
+import { BaseMathType, ChartDisplayType, PropertyFilterType, PropertyOperator } from '~/types'
+
+export function formatQueryForInsightEditor(query: InsightVizNode<TrendsQuery>): InsightVizNode<TrendsQuery> {
+    return {
+        ...query,
+        source: {
+            ...query.source,
+            dateRange: {
+                date_from: 'wStart',
+                date_to: null,
+            },
+        },
+        // Open in regular insight mode with full controls visible.
+        full: true,
+        showHeader: undefined,
+        showTable: undefined,
+        showFilters: undefined,
+        embedded: undefined,
+    }
+}
+
+export function buildExceptionVolumeQuery(dateFrom: string, dateTo: string): InsightVizNode<TrendsQuery> {
+    return {
+        kind: NodeKind.InsightVizNode,
+        source: {
+            kind: NodeKind.TrendsQuery,
+            series: [
+                {
+                    kind: NodeKind.EventsNode,
+                    event: null,
+                    custom_name: 'Exceptions',
+                    properties: [
+                        {
+                            key: 'event',
+                            type: PropertyFilterType.Event,
+                            value: '$exception',
+                            operator: PropertyOperator.Exact,
+                        },
+                    ],
+                },
+            ],
+            interval: 'day',
+            dateRange: { date_from: dateFrom, date_to: dateTo },
+            trendsFilter: { display: ChartDisplayType.ActionsBar },
+        },
+        showHeader: false,
+        showTable: false,
+    }
+}
+
+export function buildCrashFreeSessionsQuery(dateFrom: string, dateTo: string): InsightVizNode<TrendsQuery> {
+    return {
+        kind: NodeKind.InsightVizNode,
+        source: {
+            kind: NodeKind.TrendsQuery,
+            series: [
+                {
+                    kind: NodeKind.EventsNode,
+                    event: null,
+                    custom_name: 'Total sessions',
+                    math: BaseMathType.UniqueSessions,
+                },
+                {
+                    kind: NodeKind.EventsNode,
+                    event: null,
+                    custom_name: 'Sessions with crash',
+                    math: BaseMathType.UniqueSessions,
+                    properties: [
+                        {
+                            key: 'event',
+                            type: PropertyFilterType.Event,
+                            value: '$exception',
+                            operator: PropertyOperator.Exact,
+                        },
+                    ],
+                },
+            ],
+            interval: 'day',
+            dateRange: { date_from: dateFrom, date_to: dateTo },
+            trendsFilter: {
+                display: ChartDisplayType.ActionsLineGraph,
+                formulaNodes: [{ formula: '(A - B) / A * 100', custom_name: 'Crash-free sessions %' }],
+                aggregationAxisPostfix: '%',
+            },
+        },
+        showHeader: false,
+        showTable: false,
+    }
+}
+
+export function insightNewUrl(query: InsightVizNode<TrendsQuery>): string {
+    return urls.insightNew({ query: formatQueryForInsightEditor(query) })
+}
